Extract Mock compiler ctx classes from nested classes to separate package
