+++ engine/runtime-compiler/src/test/java/org/enso/compiler/test/mock/DiagnosticException.java
@@ -1,0 +1,16 @@
+package org.enso.compiler.test.mock;
+
+import org.enso.compiler.core.ir.Diagnostic;
+
+public final class DiagnosticException extends RuntimeException {
+
+  public final MockModule module;
+  public final Diagnostic diagnostic;
+  final boolean isOutputRedirected;
+
+  DiagnosticException(MockModule module, Diagnostic diagnostic, boolean isOutputRedirected) {
+    this.module = module;
+    this.diagnostic = diagnostic;
+    this.isOutputRedirected = isOutputRedirected;
+  }
+}
